Silence ConnectabilityState Warning
Bug: 305066880
Test: mma -j32
Flag: EXEMPT, compile-time fix

+++ system/gd/hci/acl_manager/le_impl.h
@@ -96,7 +96,7 @@ inline std::string connectability_state_machine_text(const ConnectabilityState& 
     CASE_RETURN_TEXT(ConnectabilityState::ARMED);
     CASE_RETURN_TEXT(ConnectabilityState::DISARMING);
     default:
-      return base::StringPrintf("UNKNOWN[%d]", state);
+      return base::StringPrintf("UNKNOWN[%d]", static_cast<int>(state));
   }
 }
